chores: reformat files for futher developement
