+++ bt_tracker_utils/TrackerEvent.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+class TrackerEvent(Enum):
+    STARTED = "started"
+    COMPLETED = "completed"
+    STOPPED = "stopped"
