+++ bt_tracker_utils/tracker/TrackerQueryException.py
